services/auth-service/requirements.txt
# Auth Service Requirements

# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy[asyncio]==2.0.23
aiosqlite==0.19.0  # For SQLite async support
asyncpg==0.29.0    # For PostgreSQL async support (optional)

# Data validation and serialization
pydantic==2.5.0
pydantic-settings==2.1.0

# JWT and Security
PyJWT==2.8.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.1
cryptography==41.0.7

# HTTP client (for inter-service communication)
httpx==0.25.2
aiohttp==3.9.1

# Date/time handling
python-dateutil==2.8.2

# Environment variables
python-dotenv==1.0.0

# Logging and monitoring
structlog==23.2.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2  # For testing API endpoints

# Development tools
black==23.11.0
isort==5.12.0
flake8==6.1.0

# Database migrations (optional - for production)
alembic==1.13.0

# Redis for caching (optional)
redis==5.0.1
aioredis==2.0.1